refactor: move util/sessions to http/sessions
Closes #1222
