+++ lib/http/preprocessors.js
@@ -14,7 +14,7 @@ const oidc = require('../util/oidc');
 const Problem = require('../util/problem');
 const { QueryOptions } = require('../util/db');
 const { reject, getOrReject } = require('../util/promise');
-const { SESSION_COOKIE } = require('../util/sessions');
+const { SESSION_COOKIE } = require('./sessions');
 
 // injects an empty/anonymous auth object into the request context.
 const emptyAuthInjector = ({ Auth }, context) => context.with({ auth: Auth.by(null) });
