+++ lib/resources/oidc.js
@@ -19,7 +19,7 @@ const { parse, render } = require('mustache');
 
 const { safeNextPathFrom } = require('../util/html');
 const { redirect } = require('../util/http');
-const { createUserSession } = require('../util/sessions');
+const { createUserSession } = require('../http/sessions');
 const { // eslint-disable-line object-curly-newline
   CODE_CHALLENGE_METHOD,
   SCOPES,
