+++ lib/resources/sessions.js
@@ -12,7 +12,7 @@ const Problem = require('../util/problem');
 const { isBlank, noargs } = require('../util/util');
 const { getOrReject, rejectIf } = require('../util/promise');
 const { success } = require('../util/http');
-const { SESSION_COOKIE, createUserSession } = require('../util/sessions');
+const { SESSION_COOKIE, createUserSession } = require('../http/sessions');
 const oidc = require('../util/oidc');
 
 module.exports = (service, endpoint) => {
